Improved thread safety in random number generation.
rand() is not thread safe, it holds internal state. A simultaneous call
may mess the internal state especially badly if the internal state consists
of more than one variable. A quick solution is to use this:

static unsigned int seed = rand();	// statics are thread safe in C++11
...
auto value = rand(&seed);

A better solution would be to use Boost.Math but for now this is enough.